update: yt-dlp-web-console
Update monkey patch to 'pyodide_http._urllib.urlopen'.

By happenstance, this patch provides the perfect opportunity
to update URLs that would otherwise be blocked for mixed content
before the request is made with Javascript XHR.

The issue is that extractor scripts aren't written
with browser security policies in mind,
because they are intended to be run in a Node.js environment.

When the URL scheme of the page running in the browser is 'https:'
> for example:
>   https://warren-bank.github.io/single-page-apps/yt-dlp-web-console/index.pyodide-0.27.0-fork-2.0.0.html

and the extractor requests a URL having an insecure scheme
> for example, the NBC extractor:
>   https://github.com/yt-dlp/yt-dlp/blob/2024.12.23/yt_dlp/extractor/nbc.py#L215
> requests the URL:
>   http://link.theplatform.com/s/NnzsPC/media/guid/...

then the XHR request will be blocked,
unless the scheme for this URL were silently upgraded to 'https:'.